changed value to int and created counter that displays the total price of the order

+++ app/src/main/java/classes/OrderItem.kt
@@ -2,8 +2,8 @@ package classes
 
 class OrderItem (var restaurantName : String,
                  var orderFromMeny : String,
-                 var price : Double,
-                 var deliveryFee : Double){
+                 var price : Int,
+                 var deliveryFee : Int){
